add secret token
