+++ server/routes.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-router.get("/", (req, res) => {
+router.get("/", (_req, res) => {
   res.send({ response: "Server up" }).status(200);
 });
 module.exports = router;
